Fixed calling callback

+++ src/android/SegmentCordovaPlugin.java
@@ -324,7 +324,7 @@ public class SegmentCordovaPlugin extends CordovaPlugin {
 
             isInitialized = analytics != null;
 
-            callbackContext.success(isInitialized);
+            callbackContext.success((int) isInitialized);
         } catch (Exception ex) {
             Log.getStackTraceString(ex);
         }
